Refactor and tweak load module.
