--- src/components/graph/Menu.css
@@ -1,2 +1,0 @@
-.Menu {
-}
