--- src/components/graph/Menu.js
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-
-import './Menu.css';
-
-import type { Module } from './types/Module';
-
-export default function Menu(props) {
-  return (
-  );
-}
